Move Combobox to its own folder
