+++ src/components/Combobox/index.js
@@ -1,0 +1,13 @@
+import Combobox from './Combobox';
+import ControlledCombobox from './ControlledCombobox';
+import ComboboxDropdown from './ComboboxDropdown';
+import ComboboxDropdownList from './ComboboxDropdownList';
+import ComboboxDropdownListItem from './ComboboxDropdownListItem';
+
+export {
+  Combobox,
+  ControlledCombobox,
+  ComboboxDropdown,
+  ComboboxDropdownList,
+  ComboboxDropdownListItem,
+};
